# Fix memory access in BeamBins