--- a/src/particles/BeamParticleContainer.H
+++ b/src/particles/BeamParticleContainer.H
@@ -26,7 +26,29 @@ struct BeamIdx
     };
 };
 
-using BeamBins = amrex::DenseBins<amrex::ParticleTile<0, 0, BeamIdx::nattribs, 0>::ParticleType>;
+struct BeamBins : amrex::DenseBins<amrex::ParticleTile<0, 0, BeamIdx::nattribs, 0>::ParticleType> {
+
+    template<class...Args>
+    void build (Args&&...args) {
+        amrex::DenseBins<amrex::ParticleTile<0, 0, BeamIdx::nattribs, 0>::ParticleType>::build(args...);
+
+        const auto offset_size = numBins() + 1;
+        const auto offsets_gpu = offsetsPtr();
+        m_offsets_cpu.resize(offset_size);
+#ifdef AMREX_USE_GPU
+        amrex::Gpu::dtoh_memcpy_async(m_offsets_cpu.dataPtr(), offsets_gpu,
+                                      offset_size * sizeof(index_type));
+        amrex::Gpu::streamSynchronize();
+#else
+        std::memcpy(m_offsets_cpu.dataPtr(), offsets_gpu,
+                    offset_size * sizeof(index_type));
+#endif
+    }
+
+    const index_type* offsetsPtrCpu () const noexcept { return m_offsets_cpu.dataPtr(); }
+
+    amrex::Vector<index_type> m_offsets_cpu;
+};
 
 /** \brief Container for particles of 1 beam species. */
 class BeamParticleContainer