This is the same issue as in #798, however here the performance is not affected much with the current memory page layout. This PR is necessary to avoid segfaults with the option
```
amrex.the_arena_is_managed = 0
```

Runtime in s:

resolution | old | new
-- | -- | --
16 | 0.5078 | 0.5039
32 | 0.8205 | 0.7247
64 | 1.5510 | 1.4600
128 | 1.9580 | 1.9480
256 | 6.9270 | 6.8080
512 | 19.5200 | 19.5100
1024 | 69.9800 | 69.9400


- [ ] **Small enough** (< few 100s of lines), otherwise it should probably be split into smaller PRs
- [ ] **Tested** (describe the tests in the PR description)
- [ ] **Runs on GPU** (basic: the code compiles and run well with the new module)
- [ ] **Contains an automated test** (checksum and/or comparison with theory)
- [ ] **Documented**: all elements (classes and their members, functions, namespaces, etc.) are documented
- [ ] **Constified** (All that can be `const` is `const`)
- [ ] **Code is clean** (no unwanted comments, )
- [ ] **Style and code conventions** are respected at the bottom of https://github.com/Hi-PACE/hipace
- [ ] **Proper label and GitHub project**, if applicable